rename: router news로 이름 변경 및 뉴스콘텐츠에 제목 추가
