+++ data/news.json
@@ -26,7 +26,7 @@
     },
     {
       "title": "와이바이오로직스, 美 AACR서 항-CD39 항체의 면역항암 효능 개선 연구 결과 발표",
-      "description": "기존 경쟁 약물 대비 CD39에 대한 높은 결합능력과 저해 효과로 ATP의 전환 억제 확인\n주 2회 투약한 동물실험에서도 종양의 성장 억제 효과와 우수한 내약성\n향후 면역항암제 병용투여에서 면역항암제 치료 효과 개선에 시너지 효과 기대\n종양미세환경에서의 ATP 전환에 의한 면역세포 활성 억제 기전(사진=와이바이로직스)\n\n[바이오타임즈] 항체 신약 개발 플랫폼 기업 와이바이오로직스(각자대표 박영우·장우익, KOSDAQ 338840)는 2024년 미국암연구학회 연례회의(AACR Annual Meeting 2024)에서 항-CD39 항체 ‘AR062’의 종양미세환경(TME, Tumor Microenvironment) 개선을 통한 면역항암 효능에 대한 연구 결과를 포스터 형식으로 공개했다고 9일 밝혔다.\n\n와이바이오로직스는 이번 행사에서 항체 라이브러리인 Ymax®-ABL(와이맥스에이블)을 통해 발굴한 항-CD39 항체 ‘AR062’의 연구 결과를 포스터 형식으로 발표했다. AR062는 CD39를 저해하여 종양미세환경(TME)을 개선함으로써 면역세포 활성화를 유도해 면역항암 작용을 일으키며, 이를 바탕으로 면역항암제와의 병용투여에서 시너지 효과를 기대할 수 있다는 게 회사의 설명이다.\n\nCD39는 세포막 외부에 위치해 세포 밖에 존재하는 아데노신삼인산(ATP)을 아데노신(Adenosine)으로 분해하는 과정에 참여한다. 이 과정을 통해서 세포 밖에 축적된 아데노신은 면역세포의 활성을 억제하는 종양미세환경을 조성하게 된다.\n \n\nAR062에 의한 면역항암 효과 기전 모식도(사진=와이바이로직스)\n\n회사의 연구 결과에 따르면, AR062는 항-CD39 단일클론항체로, CD39가 ATP를 아데노신으로 변환하는 것을 억제하는 동시에 면역세포 활성을 돕는 ATP를 축적하게 만든다. 아데노신이 면역세포 활성을 억제하는 작용이 있기 때문에, 이러한 변환 과정의 억제는 종양 주변에서의 면역세포 활성을 높일 수 있고, 이를 통해 면역항암제의 체내 효능을 극대화할 수 있다는 의미다.\n\n와이바이오로직스의 이번 발표에 따르면, AR062는 기존 경쟁 약물 대비 CD39에 대한 높은 결합능력과 저해 효과를 보이며, 강력한 결합력을 바탕으로 ATP의 전환을 억제함을 확인했다. 또한, 주 2회 투약한 동물실험에서도 종양의 성장 억제 효과와 우수한 내약성 또한 확인했다."
+      "description": "기존 경쟁 약물 대비 CD39에 대한 높은 결합능력과 저해 효과로 ATP의 전환 억제 확인\n주 2회 투약한 동물실험에서도 종양의 성장 억제 효과와 우수한 내약성\n향후 면역항암제 병용투여에서 면역항암제 치료 효과 개선에 시너지 효과 기대\n종양미세환경에서의 ATP 전환에 의한 면역세포 활성 억제 기전(사진=와이바이로직스)\n\n[바이오타임즈] 항체 신약 개발 플랫폼 기업 와이바이오로직스(각자대표 박영우·장우익, KOSDAQ 338840)는 2024년 미국암연구학회 연례회의(AACR Annual Meeting 2024)에서 항-CD39 항체 ‘AR062’의 종양미세환경(TME, Tumor Microenvironment) 개선을 통한 면역항암 효능에 대한 연구 결과를 포스터 형식으로 공개했다고 9일 밝혔다.\n\n와이바이오로직스는 이번 행사에서 항체 라이브러리인 Ymax®-ABL(와이맥스에이블)을 통해 발굴한 항-CD39 항체 ‘AR062’의 연구 결과를 포스터 형식으로 발표했다. AR062는 CD39를 저해하여 종양미세환경(TME)을 개선함으로써 면역세포 활성화를 유도해 면역항암 작용을 일으키며, 이를 바탕으로 면역항암제와의 병용투여에서 시너지 효과를 기대할 수 있다는 게 회사의 설명이다.\n\nCD39는 세포막 외부에 위치해 세포 밖에 존재하는 아데노신삼인산(ATP)을 아데노신(Adenosine)으로 분해하는 과정에 참여한다. 이 과정을 통해서 세포 밖에 축적된 아데노신은 면역세포의 활성을 억제하는 종양미세환경을 조성하게 된다.\n \n\nAR062에 의한 면역항암 효과 기전 모식도(사진=와이바이로직스)"
     },
     {
       "title": "'황선홍 감독 노력에도...' 양현준→김지수 합류 불발... '잔류 경쟁' 소속팀 차출불가 통보→포항 김동진 대체발탁  머니투데이",
@@ -46,7 +46,7 @@
     },
     {
       "title": "폐경 전 '홍조, 발한' 등 갱년기 증상 있었다면 우울증 더욱 심할 수 있어",
-      "description": "[메디칼업저버 박선재 기자] 폐경 전 안면홍조 및 야간발한 증상을 경험한 갱년기 여성은 우울 증상을 겪을 위험이 더욱 높다는 연구 결과가 발표됐다.\n\n강북삼성병원 헬스케어데이터센터 류승호·장유수 교수, 최혜린 박사 연구팀은 2014년~2018년 사이 강북삼성병원 종합건진센터를 방문한 42세~52세의 폐경 전 갱년기 여성 2,800여 명을 대상으로 평균 6.1년의 추적 분석을 실시했다.\n\n연구팀은 안면홍조 및 야간발한 증상은 설문조사를 통해 측정했으며, 증상이 있다고 응답한 사람에서 1점-7점까지 척도 중 3점 이상의 괴로움을 느끼는 경우 중등도 이상의 혈관운동증상이 있다고 정의했다.\n\n우울 증상은 우울 설문조사를 통해 지난 일주일간의 우울 수준을 조사했으며, 총점 60점 중 16점 이상이면 유의미한 우울 증상이 있는 것으로 간주했다.\n\n연구팀은 우울 증상이 없는 그룹을 6년간 추적 조사를 실시했다.\n\n3점 이상 중등도 혈관운동 증상 겪은 여성은 우울 증상 83%까지 증가 \n\n그 결과, 혈관운동 증상을 경험한 여성에서 그렇지 않은 여성에 비해 우울 증상 발생 위험이 약 72% 더 높은 것으로 나타났다. 특히 3점 이상의 중등도 혈관운동 증상을 겪은 여성은 혈관운동 증상이 없는 여성에 비해 우울 증상 발생 위험이 약 83%까지 증가하는 것으로 확인되었다.\n\n류승호 교수는 “갱년기 여성은 여성호르몬의 불균형에 의해 혈관운동성 증상을 경험하게 되는데, 폐경 전에 중증도의 갱년기 증상을 경험하는 여성인 경우 불균형적인 여성호르몬의 변화에 유의한 우울증상에 더 취약할 수 있다.”며 상관관계에 대해 추정했다.\n\n이어 “폐경 전에 안면홍조 및 야간발한을 겪는 갱년기 여성에서 홍조 자체의 증상 뿐 아니라, 정신건강을 포함한 포괄적인 관리가 중요하다는 것을 시사한다”며 연구의 의미를 강조했다.\n\n장유수 교수는 “전향적 갱년기 여성 연구를 통해 갱년기 증상과 만성 질환의 다양한 위험 요인을 밝히고, 이를 기반으로 중년 여성의 신체적 및 정신적 삶의 질 향상에 기여하기 위해 지속적으로 노력할 것”이라고 밝혔다.\n\n한편 이번 연구는 국제 정신질환학회의 공식 학술지인 정신질환 저널(Journal of Affective Disorders) 에 게재되었으며, 질병관리청 국립보건연구원의 갱년기 여성의 만성질환 위험 요인 규명을 위한 전향적 연구 사업의 지원으로 수행되었다."
+      "description": "[메디칼업저버 박선재 기자] 폐경 전 안면홍조 및 야간발한 증상을 경험한 갱년기 여성은 우울 증상을 겪을 위험이 더욱 높다는 연구 결과가 발표됐다.\n\n강북삼성병원 헬스케어데이터센터 류승호·장유수 교수, 최혜린 박사 연구팀은 2014년~2018년 사이 강북삼성병원 종합건진센터를 방문한 42세~52세의 폐경 전 갱년기 여성 2,800여 명을 대상으로 평균 6.1년의 추적 분석을 실시했다.\n\n연구팀은 안면홍조 및 야간발한 증상은 설문조사를 통해 측정했으며, 증상이 있다고 응답한 사람에서 1점-7점까지 척도 중 3점 이상의 괴로움을 느끼는 경우 중등도 이상의 혈관운동증상이 있다고 정의했다.\n\n우울 증상은 우울 설문조사를 통해 지난 일주일간의 우울 수준을 조사했으며, 총점 60점 중 16점 이상이면 유의미한 우울 증상이 있는 것으로 간주했다.\n\n연구팀은 우울 증상이 없는 그룹을 6년간 추적 조사를 실시했다.\n\n3점 이상 중등도 혈관운동 증상 겪은 여성은 우울 증상 83%까지 증가 \n\n그 결과, 혈관운동 증상을 경험한 여성에서 그렇지 않은 여성에 비해 우울 증상 발생 위험이 약 72% 더 높은 것으로 나타났다. 특히 3점 이상의 중등도 혈관운동 증상을 겪은 여성은 혈관운동 증상이 없는 여성에 비해 우울 증상 발생 위험이 약 83%까지 증가하는 것으로 확인되었다.\n\n류승호 교수는 “갱년기 여성은 여성호르몬의 불균형에 의해 혈관운동성 증상을 경험하게 되는데, 폐경 전에 중증도의 갱년기 증상을 경험하는 여성인 경우 불균형적인 여성호르몬의 변화에 유의한 우울증상에 더 취약할 수 있다.”며 상관관계에 대해 추정했다.\n\n이어 “폐경 전에 안면홍조 및 야간발한을 겪는 갱년기 여성에서 홍조 자체의 증상 뿐 아니라, 정신건강을 포함한 포괄적인 관리가 중요하다는 것을 시사한다”며 연구의 의미를 강조했다."
     },
     {
       "title": "전립선암 판별 바로미터 PSA 검사 효용성 논란 지속",
